refactor: update dependencies and enhance upload handling
- Updated `zod` version from 4.5.2 to 4.5.4 for improved validation features.
- Upgraded `lucide-react` from 1.35.0 to 1.37.0 to incorporate the latest icon updates.
- Revised `tsx` version from 4.23.12 to 4.23.13 for better TypeScript support.
- Enhanced upload handling by ensuring file size validation is strictly enforced as an integer.
- Improved documentation to clarify changes in upload policies and error handling.
- Updated tests to validate new dependency versions and upload behavior.

+++ packages/core/src/schema/multipart.ts
@@ -9,7 +9,7 @@ export const multipartSignPartBodySchema = z.object({
   key: trimmedString,
   uploadId: trimmedString,
   partNumber: z.number().int().positive(),
-  partSize: z.number().positive(),
+  partSize: z.number().int().positive(),
 });
 
 export const multipartListPartsQuerySchema = z.object({
